Add `@ts-nocheck` to `src/data/tms.ts`, move `tmPoolTiers` to own file
