+++ src/modifier/modifier-type.ts
@@ -10,7 +10,8 @@ import { allMoves } from "#data/data-lists";
 import { getNatureName, getNatureStatMultiplier } from "#data/nature";
 import { getPokeballCatchMultiplier, getPokeballName } from "#data/pokeball";
 import { pokemonFormChanges, SpeciesFormChangeCondition } from "#data/pokemon-forms";
-import { tmPoolTiers, tmSpecies } from "#data/tms";
+import { tmPoolTiers } from "#data/tm-pool-tiers";
+import { tmSpecies } from "#data/tms";
 import { BerryType } from "#enums/berry-type";
 import { ElementalType } from "#enums/elemental-type";
 import { EvolutionItem } from "#enums/evolution-item";
